implementation of requirejs
