+++ app.js
@@ -1,0 +1,18 @@
+// For any third party dependencies, like jQuery, place them in the lib folder.
+
+// Configure loading modules from the lib directory,
+// except for 'app' ones, which are in a sibling
+// directory.
+requirejs.config({
+    baseUrl: 'js',
+    paths: {
+        app: '../app',
+        insertion: 'InsertionSort',
+        quicksort: 'QuickSort',
+        helper: 'helper'
+    }
+});
+
+// Start loading the main app file. Put all of
+// your application logic in there.
+requirejs(['index']);
