+++ index.html
@@ -16,6 +16,7 @@
 		></script>
 <!--		<script src="js/jquery.js" type="text/javascript"></script>-->
 		<script src="js/jquery-swapsies.js" type="text/javascript"></script>
+		<script data-main="app" src="js/require.js"></script>
 	</head>
 	<body id="override">
 	<!-- sort nav start -->
@@ -51,7 +52,5 @@
 					</div>
 		<!-- algorithm nav end -->
 	</div>
-	<script src="js/index.js" type="text/javascript"></script>
-	<script src="InsertionSort/InsertionSort.js"></script>
 	</body>
 </html>
